pyproject.toml
[tool.poetry]
name = "sage"
version = "0.0.9"
description = "Sage: A conversational AI assistant simplifying data interactions with intuitive ease"
authors = ["Ayo Ayibiowu <charles.ayibiowu@hotmail.com>"]
license = "Apache-2.0 license"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
anyio = "^3.7.1"
asyncer = "^0.0.2"
chainlit = "1.1.300rc4"
cohere = "4.40"
duckduckgo-search = "4.1.1"
aiohttp = "^3.9.1"
faiss-cpu = "1.7.4"
gitpython = "^3.1.43"
html2text = "^2020.1.16"
huggingface-hub = "0.20.2"
jira = "^3.8.0"
langchain = "^0.1.12"
langchain-community = "^0.0.28"
langchain-core = "^0.1.31"
markdownify = "0.11.6"
openai = "^1.13.3"
pdf2image = "1.17.0"
pydantic = "2.5.3"
python-docx = "1.1.2"
pyyaml = "6.0.1"
requests = "2.31.0"
sentence-transformers = "^2.7.0"
toml = "0.10.2"
transformers = "4.36.2"
unstructured = { extras = ["docx", "pdf", "xlsx"], version = "^0.12.5" }
xlsxwriter = "3.1.9"
python-gitlab = "4.3.0"
markdown = "^3.5.2"
apscheduler = "^3.10.4"
croniter = "^2.0.2"
atlassian-python-api = "^3.41.11"
aiosqlite = "^0.20.0"
litellm = "^1.35.23"
dataclasses-json = "0.5.14"
crewai = "^0.30.11"

[tool.poetry.group.tests]
optional = true

[tool.poetry.group.tests.dependencies]
pytest = "^8.2.2"
pytest-cov = "^4.1.0"

[tool.poetry.group.dev.dependencies]
codespell = { version = "^2.2.6" }
ruff = { version = "^0.3.0" }

[tool.pytest.ini_options]
testpaths = ["tests/unit_tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    "default::DeprecationWarning:sage.*",
    "default::PendingDeprecationWarning:sage.*",
]

[tool.codespell]
skip = "./.git,./data"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
